Refactored React components and modularized their CSS
