+++ client/main.scss
@@ -34,9 +34,6 @@ $_next-ui-setup-first-pass: false;
 
 @import "icons";
 
-@import "../components/feed/main";
-@import "../components/fastft/main";
-
 @import "components/flexipod/main";
 
 @import "flexipod-layouts";
@@ -239,15 +236,6 @@ body {
 			display: block;
 		}
 	}
-}
-
-#fastft {
-	h1 {@include nhFastFtTitle; }
-	article { @include nhFeedItem(fastft-brand); }
-}
-
-.feed {
-	article { @include nhFeedItem(link-2); }
 }
 
 .bordered-spacer {
