+++ components/feed/main.scss
@@ -1,5 +1,4 @@
-@import "next-type/main";
-@import "o-colors/main";
+@import "next-sass-setup/main";
 
 @mixin nhFeedItem($hover-color) {
 	border: 1px oColorsGetPaletteColor(warm-2) solid;
@@ -69,4 +68,8 @@
 .fade-leave.fade-leave-active {
   opacity: 0.01;
   transition: opacity 0.5s ease-in;
+}
+
+.feed {
+	article { @include nhFeedItem(link-2); }
 }
